More efficient entry reads and better default entry support in DeviceMgr
 * ability to read a single match entry through PI instead of always
   having to read the entire table
 * ability to read a single member / group instead of always having to
   read all members + all groups
 * ability to read default entry
 * ability to configure default resources for default entry

Fixes #399
Fixes #400
Fixes #401

New PI functions:
 * pi_table_default_action_get_handle to retrieve the entry handle of
   the default entry; we assume this handle always exists (one
   pi_update_device_start returns) and never changes (for a given table)
   during the lifetime of the P4 forwarding pipeline config.
 * pi_table_entries_fetch_one to read a single match entry (from handle)
 * pi_table_entries_fetch_wkey to read a single match entry (from match
   key)
 * pi_act_prof_mbr_fetch to retrieve a single member
 * pi_act_prof_grp_fetch to retrieve a single group

We are currently not assuming that PI table functions, which take the
entry handle (e.g. pi_table_entries_fetch_one) or the match key
(e.g. pi_table_entries_fetch_wkey) as a parameter, work for the default
entry (when passing NULL as the match key or the default handle as the
entry handle), but that may change in the future (and we may deprecate
the current PI functions which specifically exist for the default
entry). However, read / write methods for direct resources'
configuration are assumed to work when called with the default entry
handle.