Add runtime check for ABI compatibility
This ensures that the core PI library and the target-specific PI
implementation are "ABI-compatible". Or in other words it ensures that
the PI library which was used to build the target-specific PI
implementation is the "same" as the PI library which ultimately is
linked with the target-specific PI implementation. The purpose of this
is to avoid some runtime issues / crashes which can be hard to debug.

This is a poor man workaround. Ideally we could try to rely on something
like library SONAME, but at the moment targets are free to provide their
PI implementation in any form they want, including as part of a larger
SDK library.

Fixes #448