Support reads for PRE entries
We support reads for both multicast groups and clone sessions. We
"cheat" and take a slightly different approach than usual: instead of
calling into PI, we instead rely entirely on local state stored in the
DeviceMgr to reply to the ReadRequests. The P4Runtime spec does not
mandate anything in this regard and does not require Read RPCs to read
state from HW or lower-level drivers. It is up to the implementation,
and while we usually choose to read the state from PI, here we chose not
to, in order to reduce the amount of work required to implement the
feature.

Note that the current implementation does not return the "replicas" list
in the same order as in the corresponding WriteRequest, which is allowed
by the P4Runtime spec.

Fixes #492