Update P4Runtime version to 1.1.0-rc.1 and implement Capabilities RPC
As of this version of P4Runtime, the Capabilities RPC only returns the
version of the API implemented by the server.

The RPC is handled entirely by the server, and DeviceMgr is not
involved. This means that if the DeviceMgr implementation
(libpifecpp.so) is updated separately from the server, the version
number returned by the server may not be meaningful anymore. However,
moving the implementation of Capabilities to DeviceMgr does not really
resolve the issue either, since some parts of P4Runtime (such as
master-arbitration) are the responsibility of the server. The current
solution is consistent with Stratum, which takes care of implementing
Capabilities in the target-independent server code.