Add PIGrpcServerInit to P4Runtime server C API
Without this change, the caller of PIGrpcServerRun is required to call
DeviceMgr::init directly, which does not make much sense since the API
is meant to be usable by C code to start a P4Runtime gRPC server.

We also call DeviceMgr::destroy unconditionally in PIGrpcServerCleanup.

This change will not break current users of the API (who currently call
DeviceMgr::init directly) - which was the reason why we decided not to
call DeviceMgr::init as part of the PIGrpcServerRun definition. To
achieve this, we made sure that having more than one call to
DeviceMgr::destroy was harmless. We still recommend current users to
update their code.

The PI_ABI_VERSION number is increased to 3 because we added
PI_STATUS_INIT_ALREADY_CALLED to the pi_status_t enum.

Fixes #512