Fix deadlock in P4Runtime server implementation
It was observed that if a P4Runtime client was reading messages from the
stream channel and reacting to them by issuing Write RPCs in the same
thread, a deadlock was possible in the server implementation.

We address this by changing the synchronization scheme for the
DeviceState class. A unit test is added to validate the fix. We
confirmed that the test was failing before applying the patch.

Fixes #523