Remove ppa:ubuntu-toolchain-r/test from Dockerfiles
For some reason, having this PPA makes it impossible to install clang,
it creates a "dependency hell". This is recent a issue and wasn't
present a couple weeks ago. To avoid the issue altogether, rather than
waste a lot of time debugging it, we remove the PPA, which means we no
longer run a CI job with a "recent" version of gcc (gcc 7). We still
test with the default version of gcc for Ubuntu 16.04, along with
clang-3.8 and clang-8 (updated from clang-6.0).

The long-term solution should really be to update the base image from
Ubuntu 16.04 to 20.04 (https://github.com/p4lang/third-party/issues/25).