Unsafe proto release cleanup (#546)
MUST_USE_RESULT is being added to these release methods, which will break this code as it's currently written. This change is part of an internal google effort to fix these instances before enabling PROTOBUF_MUST_USE_RESULT.

The unsafe arena variants of set_allocated_* and release_* will not require the result to be checked, and will also avoid unwanted copies or memory leaks if the protos are owned by an arena.