Fix serialization of PacketOut metadata headers
The server code was supposed to handle metadata fields provided
out-of-order, but the serialization function itself (generic_deparse(),
taken from the bmv2 code), assumed that it was called in the order in
which the fields were defined in the controller header. A call to
generic_deparse() may cause existing data in the destination buffer to
be overwriiten, more precisely the data which is located in the last
byte of the buffer that needs to be written to.

For now, we add a first pass on the PacketOut metadata to order the
fields correctly. In the future, we may want to update generic_deparse()
to lift the current restriction.

Fixes #554