Remove dependency on libJudy (#561)
We replace all occurrences of Judy arrays with other data structures.
The rationale for this is that we have been running into issues with
some build systems; see
https://github.com/p4lang/behavioral-model/issues/1001.
Additionally, libJudy is not actively maintained.

We selected uthash as a replacement when a hash map is needed. The
uthash source code is copied directly under third_party/, so no
additional dependency is introduced for PI users.