configure: Update AM_PATH_PYTHON to Python 3 (#583)
* configure: Update AM_PATH_PYTHON to Python 3

The AM_PATH_PYTHON macro will search for a Python interpreter on the
system. The first argument is the minimum version of Python required.
Usually, when AM_PATH_PYTHON is invoked with an argument of the form
3.X, it means that only a Python 3 interpreter should be used. However,
the "python" command may invoke a Python 2 interpreter[1] on some
systems. Thus, if we try "python" first we may get Python 2 interpreter.

[1] http://www.python.org/dev/peps/pep-0394/

Signed-off-by: Radostin Stoyanov <rstoyanov@fedoraproject.org>

* Use python3 interpreter

The use of the "python" command is not recommended due to confusion
with references to Python 2. The recommendation is to write
cross-platform scripts with shebangs for either python2 or python3.

https://peps.python.org/pep-0394/

Signed-off-by: Radostin Stoyanov <rstoyanov@fedoraproject.org>